requirements-nopandas.txt
Flask==3.1.0
gunicorn==23.0.0
cryptography==43.0.3
azure-functions==1.20.0
pyodbc==5.2.0
requests==2.32.3
python-dotenv==1.1.1
APScheduler==3.10.4
openpyxl==3.1.5
PyJWT==2.9.0
bcrypt==4.2.0
Flask-Cors==4.0.1
